backend/requirements.txt
Flask
docling
pypdf
Flask-Cors
pdf2image
openai
Pillow
python-dotenv